+++ app/JsonApi/Admin/V1/PartnerUsers/PartnerUserRequest.php
@@ -13,6 +13,7 @@ class PartnerUserRequest extends ResourceRequest
     {
 
         if($this->isMethod("DELETE")) return;
+        if($this->isMethod('PATCH')) return;
 
         $data = $this->data;
         $data['attributes']['phone'] = preg_replace('/[^0-9]/', '', $this->data['attributes']['phone']);
@@ -34,11 +35,16 @@ class PartnerUserRequest extends ResourceRequest
             $unique = $unique->ignore($partner_user);
         }
 
-        return [
+        $rules =  [
             'partnerId' => 'required|exists:partners,id',
             'phone' => 'required|digits_between:3,15|' . $unique,
-            "status" => 'boolean'
+            'status' => 'boolean'
         ];
+        if($this->isMethod('PATCH')){
+            unset($rules['phone']); // убрать проверку на номер телефона при редактировании пользователя
+            unset($rules['partnerId']); // убрать проверку на companyId, т.к. компанию после назначения уже поменять нельзя
+        }
+        return $rules;
     }
 
 }